kernel+mem+riscv: SMP support: serialization and locks

+++ src/tests/kernel/thread/sync/concurrent_mutex_test.c
@@ -6,6 +6,8 @@
  * @author Vita Loginova
  */
 
+#include <stdint.h>
+
 #include <embox/test.h>
 #include <kernel/thread/sync/mutex.h>
 #include <kernel/thread.h>
@@ -26,7 +28,9 @@ TEST_CASE("General") {
 }
 
 static void *run(void *arg) {
-	if ((int)arg == 1)
+	/* Through intptr_t, or clang refuses the narrowing cast
+	 * outright (-Wvoid-pointer-to-int-cast is an error in this tree). */
+	if ((intptr_t)arg == 1)
 	{
 	test_assert_zero(thread_launch(t2));
 	}
